+++ core/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/assetconnection/AssetConnectionManager.java
@@ -89,19 +89,7 @@ public class AssetConnectionManager {
             LOGGER.info("Connecting to assets...");
         }
         for (var connection: connections) {
-            try {
-                // try to connect in synchronized way, if that fails keep trying to connect async
-                tryConnecting(connection);
-                setupSubscriptions(connection);
-            }
-            catch (AssetConnectionException e) {
-                LOGGER.info(
-                        "Establishing asset connection failed on initial attempt (endpoint: {}). Connecting will be retried every {}ms but no more messages about failures will be shown.",
-                        connection.getEndpointInformation(),
-                        coreConfig.getAssetConnectionRetryInterval(),
-                        e);
-                setupConnectionAsync(connection);
-            }
+            setupConnectionAsync(connection);
         }
         lambdaAssetConnection.start();
     }
@@ -178,21 +166,30 @@ public class AssetConnectionManager {
 
 
     private void tryConnectingUntilSuccess(AssetConnection connection) {
+        try {
+            tryConnecting(connection);
+        }
+        catch (AssetConnectionException e) {
+            LOGGER.info(
+                    "Establishing asset connection failed on initial attempt (endpoint: {}, reason: {}). Connecting will be retried every {} ms but no more messages about failures will be shown.",
+                    connection.getEndpointInformation(),
+                    e.getMessage(),
+                    coreConfig.getAssetConnectionRetryInterval(),
+                    e);
+        }
         while (active && !connection.isConnected()) {
             try {
                 tryConnecting(connection);
             }
             catch (AssetConnectionException e) {
+                LOGGER.trace("Establishing asset connection failed (endpoint: {})",
+                        connection.getEndpointInformation(),
+                        e);
                 try {
-                    LOGGER.trace("Establishing asset connection failed (endpoint: {})",
-                            connection.getEndpointInformation(),
-                            e);
-
                     Thread.sleep(coreConfig.getAssetConnectionRetryInterval());
                 }
                 catch (InterruptedException e2) {
-                    LOGGER.error("Error while establishing asset connection (endpoint: {})", connection.getEndpointInformation(), e2);
-                    Thread.currentThread().interrupt();
+                    // intentionally empty
                 }
             }
         }